Add usage to function status response
* Allow the function to return its CPU and RAM allocation
* Update Go version for CI to 1.17
