+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/openfaas/faas-provider
 
-go 1.13
+go 1.16
 
 require (
 	github.com/gorilla/mux v1.8.0
